20/7
crud login register logout

+++ database/migrations/2023_07_18_022618_create_news_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title');
             $table->string('description');
             $table->string('image')->nullable();
-            $table->unsignedBigInteger('news_category_id');
+            $table->integer('news_category_id');
             $table->timestamps();
         });
     }
